feat(readme): edited romfs files to move readme files from share_init to share_upgrade

+++ package/recalbox-romfs2/scripts/installroms.py
@@ -24,14 +24,15 @@ class InstallRoms:
 
     def __init__(self, systemRoot: str, target: str, root: str):
         self.__systemRoot = systemRoot
-        self.__target = target
+        self.__targetInit = target + "/share_init/roms"
+        self.__targetUpgrade = target + "/share_upgrade/roms"
         self.__tags = IniSettings(os.path.join(systemRoot, self.__TAGS_FILE), True)
         self.__tags.loadFile(True)
         self.__config = ConfigIn(root)
         self.__arch = self.__config.Arch
 
     def execute(self):
-        print("[ROMFS 2] Copying systems and metadata to {}".format(self.__target))
+        print("[ROMFS 2] Copying systems and metadata to {}".format(self.__targetInit))
 
         # Run though systems folders
         for folder in os.listdir(self.__systemRoot):
@@ -41,18 +42,19 @@ class InstallRoms:
                 if holder.CoreCount < 1:
                     continue
                 print("[ROMFS 2] Copying system {}".format(folder))
-                self.__copyRomContent(subfolder, holder.RomFolder.replace("%ROOT%", self.__target) if holder.IsPort else folder)
+                self.__copyRomContent(subfolder, holder.RomFolder.replace("%ROOT%", self.__targetInit) if holder.IsPort else folder)
+                print("[ROMFS 2] Creating readme files for system {}".format(folder))
                 self.__installReadMe(folder, holder)
 
     def __copyRomContent(self, systemFolder: str, folder: str):
         # Copy roms
         roms: str = os.path.join(systemFolder, "roms")
-        system: str = os.path.join(self.__target, folder)
+        system: str = os.path.join(self.__targetInit, folder)
         shutil.rmtree(system, ignore_errors=True)
         shutil.copytree(roms, system, dirs_exist_ok=False)
 
     def __installReadMe(self, folder: str, holder: SystemHolder):
-        targetSystemPath: str = os.path.join(self.__target, holder.RomFolder.replace("%ROOT%", self.__target) if holder.IsPort else folder)
+        targetSystemPath: str = os.path.join(self.__targetUpgrade, holder.RomFolder.replace("%ROOT%", self.__targetUpgrade) if holder.IsPort else folder)
         sourceSystemPath: str = os.path.join(self.__systemRoot, folder)
         templatePath: str = os.path.join(self.__systemRoot, ".templates/system/roms")
         for fileName, language in self.__README_FILES.items():
@@ -82,10 +84,13 @@ class InstallRoms:
             newLines.extend(self.__parse(line, holder, language))
 
         # Write back
-        filePath: str = self.__find(".readme.placeholder", destinationPath)
-        filePath: str = os.path.join(os.path.dirname(filePath), filename)
-        with open(filePath, 'w') as f:
-            f.writelines(newLines)
+        systemName: str = source.replace(self.__systemRoot, "")
+        filePath: str = os.path.join(self.__targetUpgrade, holder.RomFolder.replace("%ROOT%", self.__targetUpgrade), filename) if holder.IsPort else os.path.join(self.__targetUpgrade + systemName, filename)
+        # Create readme files only for any non-port system of if readonly is set to 0
+        if holder.IsReadOnly == 0 or holder.IsReadOnly is None:
+            os.makedirs(os.path.join(self.__targetUpgrade, holder.RomFolder.replace("%ROOT%", self.__targetUpgrade)) if holder.IsPort else os.path.join(self.__targetUpgrade + systemName), exist_ok=True)
+            with open(filePath, 'w') as f:
+                f.writelines(newLines)
 
     def __parse(self, line: str, holder: SystemHolder, language: str) -> List[str]:
         # Core loop?
